requirements.txt
python-dotenv
crewai
openai
PyPDF2